requirements.txt
tornado==6.4
opencv-python~=4.9.0.80
PyYAML~=6.0.1
numpy~=1.26.4
matplotlib~=3.8.1
pillow~=10.3.0
torch~=2.2.2